requirements.txt
fastmcp>=2.7.1
httpx>=0.24.0
nest-asyncio>=1.6.0
uvicorn>=0.24.0
python-dotenv>=1.0.0